+++ packages/web/src/views/Forms/EventForm/DateTimeSection/DatePickers/DatePickers.tsx
@@ -3,10 +3,8 @@ import React, { FC } from "react";
 import { Key } from "ts-key-enum";
 import { MONTH_DAY_YEAR } from "@core/constants/date.constants";
 import { darken } from "@core/util/color.utils";
-import {
-  dateIsValid,
-  shouldAdjustComplimentDate,
-} from "@web/common/utils/web.date.util";
+import { shouldAdjustComplimentDate } from "@web/common/utils/datetime/web.datetime.util";
+import { dateIsValid } from "@web/common/utils/web.date.util";
 import { DatePicker } from "@web/components/DatePicker/DatePicker";
 import { AlignItems } from "@web/components/Flex/styled";
 import { SetEventFormField } from "../../types";
